Adding search screen

+++ app/src/main/java/com/mujapps/jetweather/model/Sys.kt
@@ -3,7 +3,7 @@ package com.mujapps.jetweather.model
 data class Sys(
     val country: String,
     val id: Int,
-    val sunrise: Int,
-    val sunset: Int,
+    val sunrise: Long,
+    val sunset: Long,
     val type: Int
 )
